Move get_tag_dict() tests to gfa utils tests file
